Working Popup UI
There is a bug where this still triggers while in dialogue, when the player's hand is empty for treating the player, and when the player is looking at an item while holding an item. While fix soon, just pushing this in so everyone has it and I don't have a huge change set.